Got updated comments from Tom Whitman.  He also created line stipple parameters for graphic.
I added a single slot in Graphic to allow for an ascii string to be used
instead of an int.  The string will be read in and converted to hex.

Example:
   stipplePattern: "0xFFFF"

This seems more user friendly than converting to int.

Updated the gauge.epp file to put the dial in the right place, now that
the GL transform is removed.