Enhancement. Simulation.cpp timing fixes.
The range allowed for dt is (0.0 < dt <= 1000).  Setting tcRate between 0 and 1 will cause
the simulation to run at intervals greater than a second (dt = 1 / tcRate).  Using a while
loop to increment the seconds based on the newPcTvUSec value keeps the simulation from
dropping seconds from the simulation time.

pcTvUSec was set to match the computer's microsecond value.  In a constructive
run with a non-slaved clock, applying this value introduces variation in the
simulation time reported throughout the outputs.  This variation is undesirable in
constructive testing.